refactor(models): consolidate event types and order model improvements
- Move event constants to events.go (previously scattered)
- Add InventoryProduct struct for inventory operations
- Add Order.Total() method for amount calculation
- Remove order_enums.go (consolidated into order.go)
