--- internal/models/order_enums.go
@@ -1,12 +1,0 @@
-package models
-
-// OrderStatus represents the status of an order.
-type OrderStatus string
-
-const (
-	OrderPending    OrderStatus = "PENDING"
-	OrderProcessing OrderStatus = "PROCESSING"
-	OrderCompleted  OrderStatus = "COMPLETED"
-	OrderCancelled  OrderStatus = "CANCELLED"
-	OrderFailed     OrderStatus = "FAILED"
-)
